+++ src/bitset.h
@@ -13,7 +13,7 @@ class bitset {
 public:
   using value_type = bool;
 
-  using word_type = uint32_t;
+  using word_type = uint64_t;
 
   using reference = bitset_reference<word_type>;
   using const_reference = bitset_reference<const word_type>;
